requirements.txt
uvicorn==0.21.1
fastapi==0.109.1
starlette==0.35.0
pydantic==1.10.7
databases==0.7.0
SQLAlchemy==1.4.47
aiomysql~=0.2.0
httpx==0.23.0
redis~=5.2.0
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
python-multipart==0.0.6
streamlit==1.40.0
numpy~=2.0.2
pandas~=2.2.3
matplotlib~=3.9.4
seaborn~=0.13.2
requests~=2.32.3
openai~=1.65.5
gitdb~=4.0.11
h11~=0.12.0
pillow~=10.4.0
pip~=23.2.1
attrs~=24.2.0
distro~=1.9.0
wheel~=0.41.2
tornado~=6.4.1
jinja2~=3.1.4
rsa~=4.9
pyasn1~=0.6.1
ecdsa~=0.19.0
pytz~=2024.2
rich~=13.9.4
pygments~=2.18.0
toml~=0.10.2
anyio~=3.7.1
sniffio~=1.3.1
six~=1.16.0
smmap~=5.0.1
click~=8.1.7
httpcore~=0.15.0
idna~=3.10
rfc3986~=1.5.0
certifi~=2024.8.30
jiter~=0.8.2
mdurl~=0.1.2
setuptools~=68.2.0
altair~=5.4.1
jsonschema~=4.23.0
narwhals~=1.13.2
pyarrow~=18.0.0
packaging~=24.1
referencing~=0.35.1
MarkupSafe~=3.0.2
python-dateutil~=2.9.0.post0
pydeck~=0.9.1
tzdata~=2024.2
cycler~=0.12.1
urllib3~=2.2.3
PyMySQL~=1.1.1
tenacity~=9.0.0
contourpy~=1.3.0
fonttools~=4.55.3
pyparsing~=3.2.0
exceptiongroup~=1.2.2
cachetools~=5.5.0
blinker~=1.8.2
kiwisolver~=1.4.7
zipp~=3.21.0